Set up and CoC
In the "Set up" section, there is a sentence starting with "to participate in a Software Carpentry workshop, you will need access to ...". In this sentence, "Software Carpentry" is changed to "R TidyVerse". Also, I think it is good to keep that the participants should abide by Carpentries Code of Conduct in this workshop as well, but I modified this part so that it is more clear that this workshop is not a Carpentries certified (?) workshop.